+++ src/Mesher/InHouse/Square_CartesianMesh.h
@@ -245,7 +245,7 @@ public:
 		return min((double)this->Nx, (double)this->Ny) / max((double)this->Nx, (double)this->Ny);
 	}
 
-	void CoarsenMesh(CoarseningStrategy elemCoarseningStgy, FaceCoarseningStrategy faceCoarseningStgy, int coarseningFactor) override
+	void CoarsenMesh(CoarseningStrategy elemCoarseningStgy, FaceCoarseningStrategy faceCoarseningStgy, double coarseningFactor) override
 	{
 		if (elemCoarseningStgy == CoarseningStrategy::StandardCoarsening)
 		{
