+++ src/Mesher/InHouse/Square_TriangularMesh.h
@@ -341,7 +341,7 @@ public:
 		return this->Elements[0]->Regularity();
 	}
 
-	void CoarsenMesh(CoarseningStrategy elemCoarseningStgy, FaceCoarseningStrategy faceCoarseningStgy, int coarseningFactor) override
+	void CoarsenMesh(CoarseningStrategy elemCoarseningStgy, FaceCoarseningStrategy faceCoarseningStgy, double coarseningFactor) override
 	{
 		if (elemCoarseningStgy == CoarseningStrategy::StandardCoarsening)
 		{
